+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -3,7 +3,7 @@ package com.techcourse;
 import jakarta.servlet.ServletContext;
 import nextstep.mvc.DispatcherServlet;
 import nextstep.mvc.RequestMappingHandlerAdapter;
-import nextstep.mvc.controller.tobe.RequestMappingHandlerMapping;
+import nextstep.mvc.controller.RequestMappingHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
