--- mvc/src/main/java/nextstep/mvc/ControllerHandlerAdapter.java
@@ -1,24 +1,0 @@
-package nextstep.mvc;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-
-public class ControllerHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(final Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response,
-                               final Object handler) throws Exception {
-        final var controller = (Controller) handler;
-        final var logicalViewName = controller.execute(request, response);
-
-        return new ModelAndView(new JspView(logicalViewName));
-    }
-}
