+++ mvc/src/main/java/nextstep/mvc/controller/HandlerKey.java
@@ -1,8 +1,7 @@
-package nextstep.mvc.controller.tobe;
-
-import nextstep.web.support.RequestMethod;
+package nextstep.mvc.controller;
 
 import java.util.Objects;
+import nextstep.web.support.RequestMethod;
 
 public class HandlerKey {
 
@@ -24,8 +23,12 @@ public class HandlerKey {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof HandlerKey)) return false;
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof HandlerKey)) {
+            return false;
+        }
         HandlerKey that = (HandlerKey) o;
         return Objects.equals(url, that.url) && requestMethod == that.requestMethod;
     }
