+++ mvc/src/test/java/nextstep/mvc/RequestMappingHandlerAdapterTest.java
@@ -2,11 +2,8 @@ package nextstep.mvc;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import java.util.stream.Stream;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.controller.tobe.HandlerExecution;
+import nextstep.mvc.controller.HandlerExecution;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -33,17 +30,6 @@ class RequestMappingHandlerAdapterTest {
                         "RequestMappingHandlerAdapter는 HandlerExecution을 support한다",
                         new HandlerExecution(null, null),
                         true
-                ),
-                Arguments.of(
-                        "RequestMappingHandlerAdapter는 HandlerExecution 외엔 support 하지 않는다",
-                        new Controller() {
-                            @Override
-                            public String execute(final HttpServletRequest req, final HttpServletResponse res)
-                                    throws Exception {
-                                return null;
-                            }
-                        },
-                        false
                 ),
                 Arguments.of(
                         "RequestMappingHandlerAdapter는 HandlerExecution 외엔 support 하지 않는다",
